requirements.txt
Brotli==1.0.9
click==8.1.3
dash==2.4.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-defer-js-import==0.0.2
dash-html-components==2.0.0
dash-table==5.0.0
et-xmlfile==1.1.0
Flask==2.1.2
Flask-Compress==1.12
future==0.18.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2
plotly==5.8.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
scipy==1.8.1
six==1.16.0
tenacity==8.0.1
uncertainties==3.1.6
Werkzeug==2.1.2
